backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
langchain
autogen
google-generativeai
chromadb
python-dotenv
